requirements.txt
colander==1.4
iso8601==0.1.12
Jinja2==2.11.1
jsonresume-validator==0.1.5
MarkupSafe==1.1.1
pycountry==17.9.23
translationstring==1.3
